jira/requirements.txt
jira
load_dotenv
